Cleared out serpentsboxes.h
I slapped my own copyright on it, seeing as there's nothing yet worth
copyrighting.  I'll have to figure out how the S-Boxes work, but it
seems like they work the same as they did with DES.  What the public
domain documentation indicates is that they can be implemented with a
series of bitwise operations (AND|OR|XOR|NOT).  But what do those
blasted numbers mean?

git-svn-id: https://fluks.googlecode.com/svn@142 6b01645e-0d02-11de-a42d-873b86d837ff