Optimize CAST-256
In cast6.c, identifiers were fixed.  The (en/de)crypt loops were
unrolled.  The beta_q(bar) functions were changed to macros so that the
block's data are passed by reference (in effect) instead of by address.
This will speed it up by some slight factor.

h(ash/mac).cpp's create() function's switches were simplified.

Assert() is a macro if NDEBUG is defined, so that the arguments are
not evaluated.

git-svn-id: https://fluks.googlecode.com/svn@164 6b01645e-0d02-11de-a42d-873b86d837ff