Ever-so-slight improvement to the twofish interface.
I've got to find something more meaningful to do with fluks.  Maybe I
could work on keeping the decrypted key out of system memory.  There's
presently security-through-obscurity in that regard, and I'm sure some
adolescent with a high opinion of themselves would say I should be
beaten for being so absent-minded.  No, it's just not a big deal at this
point.